services/ai-worker/requirements.txt
# --- Core ---
tensorflow-cpu==2.15.0
redis
pydantic-settings
pillow

# --- ML Dependencies ---
httpx
numpy<2.0.0
opencv-python-headless
nltk
pyyaml
h5py
rouge-score
pandas

# --- AWS S3 (used when STORAGE_TYPE=s3, EKS deployment) ---
boto3
